# docs/api/openapi.yaml
...
openapi: 3.1.0
...
info:
  title: API de Finanzas Inteligentes
  version: 1.0.0-draft
  description: Contrato ejecutable derivado de CONVENCIONES.md, ENDPOINTS.md, CONTRATOS.md y TRAZABILIDAD_RF.md.
servers:
- url: https://api.finanzas.example/api/v1
  description: Producción (referencial)
- url: http://localhost:8080/api/v1
  description: Desarrollo local
security:
- bearerAuth: []
...
paths:
...
  "/movimientos-recurrentes":
    get:
      tags:
      - Movimientos recurrentes
      summary: GET /movimientos-recurrentes
      operationId: getMovimientosRecurrentes
      x-requisitos-funcionales: []
      x-etapa: 2
      x-pruebas-contrato:
      - CT-RECURRENCIAS
      security:
      - bearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/CorrelationId"
      - "$ref": "#/components/parameters/Cursor"
      - "$ref": "#/components/parameters/Limite"
      - name: estado
        in: query
        required: false
        schema:
          type: string
      - name: tipo
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Operación exitosa.
          headers:
            X-Correlation-Id:
              "$ref": "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PaginaMovimientoRecurrenteResponse"
        '400':
          "$ref": "#/components/responses/Problem400"
        '401':
          "$ref": "#/components/responses/Problem401"
        '500':
          "$ref": "#/components/responses/Problem500"
      x-capacidad: CT-RECURRENCIAS
    post:
      tags:
      - Movimientos recurrentes
      summary: POST /movimientos-recurrentes
      operationId: postMovimientosRecurrentes
      x-requisitos-funcionales: []
      x-etapa: 2
      x-pruebas-contrato:
      - CT-RECURRENCIAS
      security:
      - bearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/CorrelationId"
      - name: estado
        in: query
        required: false
        schema:
          type: string
      - name: tipo
        in: query
        required: false
        schema:
          type: string
      responses:
        '201':
          description: Operación exitosa.
          headers:
            X-Correlation-Id:
              "$ref": "#/components/headers/CorrelationId"
            Location:
              "$ref": "#/components/headers/Location"
            ETag:
              "$ref": "#/components/headers/ETag"
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/MovimientoRecurrenteResponse"
        '400':
          "$ref": "#/components/responses/Problem400"
        '401':
          "$ref": "#/components/responses/Problem401"
        '404':
          "$ref": "#/components/responses/Problem404"
        '409':
          "$ref": "#/components/responses/Problem409"
        '422':
          "$ref": "#/components/responses/Problem422"
        '500':
          "$ref": "#/components/responses/Problem500"
      x-capacidad: CT-RECURRENCIAS
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/MovimientoRecurrenteRequest"
  "/movimientos-recurrentes/{recurrenteId}":
    get:
      tags:
      - Movimientos recurrentes
      summary: GET /movimientos-recurrentes/{recurrenteId}
      operationId: getMovimientosRecurrentesByRecurrenteId
      x-requisitos-funcionales: []
      x-etapa: 2
      x-pruebas-contrato:
      - CT-RECURRENCIAS
      security:
      - bearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/CorrelationId"
      - name: recurrenteId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Operación exitosa.
          headers:
            X-Correlation-Id:
              "$ref": "#/components/headers/CorrelationId"
            ETag:
              "$ref": "#/components/headers/ETag"
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/MovimientoRecurrenteResponse"
        '401':
          "$ref": "#/components/responses/Problem401"
        '404':
          "$ref": "#/components/responses/Problem404"
        '500':
          "$ref": "#/components/responses/Problem500"
      x-capacidad: CT-RECURRENCIAS
    patch:
      tags:
      - Movimientos recurrentes
      summary: PATCH /movimientos-recurrentes/{recurrenteId}
      operationId: patchMovimientosRecurrentesByRecurrenteId
      x-requisitos-funcionales: []
      x-etapa: 2
      x-pruebas-contrato:
      - CT-RECURRENCIAS
      security:
      - bearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/CorrelationId"
      - name: recurrenteId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - "$ref": "#/components/parameters/IfMatch"
      responses:
        '200':
          description: Operación exitosa.
          headers:
            X-Correlation-Id:
              "$ref": "#/components/headers/CorrelationId"
            ETag:
              "$ref": "#/components/headers/ETag"
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/MovimientoRecurrenteResponse"
        '400':
          "$ref": "#/components/responses/Problem400"
        '401':
          "$ref": "#/components/responses/Problem401"
        '404':
          "$ref": "#/components/responses/Problem404"
        '412':
          "$ref": "#/components/responses/Problem412"
        '422':
          "$ref": "#/components/responses/Problem422"
        '500':
          "$ref": "#/components/responses/Problem500"
      x-capacidad: CT-RECURRENCIAS
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/MovimientoRecurrentePatchRequest"
    delete:
      tags:
      - Movimientos recurrentes
      summary: DELETE /movimientos-recurrentes/{recurrenteId}
      operationId: deleteMovimientosRecurrentesByRecurrenteId
      x-requisitos-funcionales: []
      x-etapa: 2
      x-pruebas-contrato:
      - CT-RECURRENCIAS
      security:
      - bearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/CorrelationId"
      - name: recurrenteId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - "$ref": "#/components/parameters/IfMatch"
      responses:
        '204':
          description: Operación completada sin cuerpo.
          headers:
            X-Correlation-Id:
              "$ref": "#/components/headers/CorrelationId"
        '401':
          "$ref": "#/components/responses/Problem401"
        '404':
          "$ref": "#/components/responses/Problem404"
        '412':
          "$ref": "#/components/responses/Problem412"
        '500':
          "$ref": "#/components/responses/Problem500"
      x-capacidad: CT-RECURRENCIAS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
        maxLength: 100
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
        pattern: ^\".+\"$
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
        maxLength: 500
    Limite:
      name: limite
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  headers:
    CorrelationId:
      description: Identificador de correlación de la solicitud.
      schema:
        type: string
    ETag:
      description: Versión opaca del recurso.
      schema:
        type: string
    Location:
      description: URI del recurso creado o aceptado.
      schema:
        type: string
        format: uri-reference
  schemas:
...
    MovimientoRecurrenteRequest:
      type: object
      additionalProperties: false
      properties:
        tipo:
          type: string
          example: gasto
        monto:
          type: integer
          format: int64
          example: 180000
          minimum: 1
        categoriaIds:
          type: array
          items:
            type: string
            example: '019b-category-...'
        cuentaId:
          type: string
          format: uuid
        descripcion:
          type: string
          example: Internet
        fechaInicio:
          type: string
          format: date
          example: '2026-07-01'
        fechaFin:
          type:
          - string
          - 'null'
        frecuencia:
          type: string
          example: mensual
        cantidadOcurrencias:
          type:
          - string
          - 'null'
      required:
      - tipo
      - monto
      - categoriaIds
      - cuentaId
      - descripcion
      - fechaInicio
      - frecuencia
    MovimientoRecurrenteResponse:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        tipo:
          type: string
          example: gasto
        monto:
          type: integer
          format: int64
          example: 180000
          minimum: 1
        categorias:
          type: array
          items:
            type: object
            additionalProperties: false
            properties:
              id:
                type: string
                format: uuid
              nombre:
                type: string
                example: Servicios
            required:
            - id
            - nombre
        cuenta:
          type: object
          additionalProperties: false
          properties:
            id:
              type: string
              format: uuid
            nombre:
              type: string
              example: Debito
          required:
          - id
          - nombre
        descripcion:
          type: string
          example: Internet
        fechaInicio:
          type: string
          format: date
          example: '2026-07-01'
        fechaFin:
          type:
          - string
          - 'null'
        frecuencia:
          type: string
          example: mensual
        cantidadOcurrencias:
          type:
          - string
          - 'null'
        ocurrenciasCompletadas:
          type: integer
          format: int64
          example: 1
        proximaEjecucion:
          type: string
          format: date
          example: '2026-08-01'
        estado:
          type: string
          example: activa
        version:
          type: integer
          format: int64
          example: 2
      required:
      - id
      - tipo
      - monto
      - categorias
      - cuenta
      - descripcion
      - fechaInicio
      - frecuencia
      - ocurrenciasCompletadas
      - proximaEjecucion
      - estado
      - version
...
    ProblemDetails:
      type: object
      additionalProperties: false
      required:
      - type
      - title
      - status
      - codigo
      - correlationId
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type:
          - string
          - 'null'
        instance:
          type:
          - string
          - 'null'
        codigo:
          type: string
          pattern: "^[a-z0-9_]+$"
        correlationId:
          type: string
        errores:
          type:
          - object
          - 'null'
          additionalProperties:
            type: array
            items:
              type: string
    Paginacion:
      type: object
      additionalProperties: false
      required:
      - siguienteCursor
      - hayMas
      - limite
      properties:
        siguienteCursor:
          type:
          - string
          - 'null'
        hayMas:
          type: boolean
        limite:
          type: integer
          minimum: 1
          maximum: 100
...
    PaginaMovimientoRecurrenteResponse:
      type: object
      additionalProperties: false
      required:
      - datos
      - paginacion
      properties:
        datos:
          type: array
          items:
            "$ref": "#/components/schemas/MovimientoRecurrenteResponse"
        paginacion:
          "$ref": "#/components/schemas/Paginacion"
...
    MovimientoRecurrentePatchRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        tipo:
          type: string
          example: gasto
        monto:
          type: integer
          format: int64
          example: 180000
          minimum: 1
        categoriaIds:
          type: array
          items:
            type: string
            example: '019b-category-...'
        cuentaId:
          type: string
          format: uuid
        descripcion:
          type: string
          example: Internet
        fechaInicio:
          type: string
          format: date
          example: '2026-07-01'
        fechaFin:
          type:
          - string
          - 'null'
        frecuencia:
          type: string
          example: mensual
        cantidadOcurrencias:
          type:
          - string
          - 'null'
...
  responses:
    Problem400:
      description: Solicitud o parámetros inválidos.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
    Problem401:
      description: Autenticación ausente o inválida.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
...
    Problem404:
      description: Recurso inexistente o no visible.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
    Problem409:
      description: Conflicto de estado o duplicado.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
...
    Problem412:
      description: ETag desactualizado.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
...
    Problem422:
      description: Regla de negocio no satisfecha.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
...
    Problem500:
      description: Error interno no controlado.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
...